Add more type safety to shared worker messages
