+++ client/src/app/worker/sw-auth.interfaces.ts
@@ -1,21 +1,27 @@
+import { AuthToken } from '../domain/interfaces/auth-token';
 import { WorkerMessage, WorkerMessageContent } from './interfaces';
+
+export interface AuthMessageType<T extends WorkerMessageContent> extends WorkerMessage<T> {
+    receiver: 'auth';
+}
 
 export interface AuthStopRefreshMessageContent extends WorkerMessageContent {
     action: 'stop-refresh';
 }
 
-export interface AuthStopRefreshMessage extends WorkerMessage<AuthStopRefreshMessageContent> {
-    receiver: 'auth';
-    msg: AuthStopRefreshMessageContent;
-}
+export interface AuthStopRefreshMessage extends AuthMessageType<AuthStopRefreshMessageContent> {}
 
 export interface AuthGetCurrentAuthMessageContent extends WorkerMessageContent {
     action: 'get-current-auth';
 }
 
-export interface AuthGetCurrentAuthMessage extends WorkerMessage<AuthGetCurrentAuthMessageContent> {
-    receiver: 'auth';
-    msg: AuthGetCurrentAuthMessageContent;
+export interface AuthGetCurrentAuthMessage extends AuthMessageType<AuthGetCurrentAuthMessageContent> {}
+
+export interface AuthUpdateMessageContent extends WorkerMessageContent {
+    action: 'update';
+    params: AuthToken | null;
 }
 
-export type AuthMessage = AuthStopRefreshMessage | AuthGetCurrentAuthMessage;
+export interface AuthUpdateMessage extends AuthMessageType<AuthUpdateMessageContent> {}
+
+export type AuthMessage = AuthStopRefreshMessage | AuthGetCurrentAuthMessage | AuthUpdateMessage;
