+++ client/src/app/worker/sw-icc.interfaces.ts
@@ -1,4 +1,10 @@
-import {WorkerMessage, WorkerMessageContent} from './interfaces';
+import { WorkerMessage, WorkerMessageContent } from './interfaces';
+
+export const ICC_ENDPOINT = `icc`;
+
+interface ICCMessageType<T extends WorkerMessageContent> extends WorkerMessage<T> {
+    receiver: 'icc';
+}
 
 export interface ICCDisconnectParams {
     type: string;
@@ -10,10 +16,7 @@ export interface ICCDisconnectAction extends WorkerMessageContent {
     params: ICCDisconnectParams;
 }
 
-export interface ICCDisconnectMessage extends WorkerMessage<ICCDisconnectAction> {
-    receiver: 'icc';
-    msg: ICCDisconnectAction;
-}
+export interface ICCDisconnectMessage extends ICCMessageType<ICCDisconnectAction> {}
 
 export interface ICCConnectMessageContent extends WorkerMessageContent {
     action: `connect`;
@@ -21,11 +24,8 @@ export interface ICCConnectMessageContent extends WorkerMessageContent {
         type: string;
         meetingId: number;
     };
-};
-
-export interface ICCConnectMessage extends WorkerMessage<ICCConnectMessageContent> {
-    receiver: 'icc';
-    msg: ICCConnectMessageContent;
 }
+
+export interface ICCConnectMessage extends ICCMessageType<ICCConnectMessageContent> {}
 
 export type ICCMessage = ICCConnectMessage | ICCDisconnectMessage;
